+++ vkcs/dataplatform/resource_cluster/models.go
@@ -175,12 +175,8 @@ func UpdateClusterConfigsUsers(ctx context.Context, users []clusters.ClusterConf
 	for _, u := range users {
 		updated := false
 
-		var roleValue types.String
-		if u.Role == "" {
-			roleValue = types.StringNull()
-		} else {
-			roleValue = types.StringValue(u.Role)
-		}
+		// role is a required attribute and cannot be null in state.
+		roleValue := types.StringValue(u.Role)
 
 		for i, user := range usersV {
 			if user.Username.ValueString() == u.Username {